# backend/docs/swagger.yaml
basePath: /api/v2
definitions:
...
  github_com_norman6464_FreStyle_backend_internal_domain.SessionNote:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      sessionId:
        type: integer
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
  internal_handler.errorResponse:
    properties:
      error:
        example: unauthorized
        type: string
    type: object
...
  internal_handler.sessionNoteUpsertReq:
    properties:
      content:
        type: string
    type: object
...
info:
  contact:
    name: FreStyle Engineering
    url: https://normanblog.com
  description: |-
    新卒 IT エンジニア 向け 統合 研修 プラットフォーム の REST API。
    Clean Architecture (port = usecase/repository / adapter = adapter/persistence) で 構築。
  termsOfService: https://normanblog.com/terms
  title: FreStyle Backend API
  version: "2.0"
paths:
...
  /sessions/{sessionId}/note:
    get:
      description: AI チャット セッション に 紐づく ノート (= 学習 者 が セッション ごと に 残した メモ) を 取得。 存在 し
        ない 場合 は 404。
      parameters:
      - description: AI チャット セッション ID
        in: path
        name: sessionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_norman6464_FreStyle_backend_internal_domain.SessionNote'
        "400":
          description: DB 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: 未 作成
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: セッション ノート 取得
      tags:
      - session-notes
    put:
      consumes:
      - application/json
      description: 指定 セッション の ノート を upsert。 userId は body で 受け取らず current user 固定
        (IDOR 対策)。
      parameters:
      - description: AI チャット セッション ID
        in: path
        name: sessionId
        required: true
        type: integer
      - description: content (userId は current user 固定 / body で 受け取らない)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.sessionNoteUpsertReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_norman6464_FreStyle_backend_internal_domain.SessionNote'
        "400":
          description: バリデーション or DB 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: セッション ノート 作成 / 更新
      tags:
      - session-notes
...
securityDefinitions:
  CookieAuth:
    description: Cognito 由来 の JWT を HttpOnly Cookie で 送る。 ログイン 後 自動 付与。
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"